Foundation Cron

+++ CINS-Server-Express/App/Controllers/UsersController.js
@@ -1,8 +1,9 @@
 const UsersModel = require('../Models/UsersModel');
 
 class UsersController {
-    setUserSession(req) {
+    setUserSession(req, userId) {
         req.session.userLoggedIn = true;
+        req.session.userId = userId;
     }
 
     setUserCookie(res, userId) {
@@ -41,14 +42,66 @@ class UsersController {
         return getUser;
     }
 
-    async postUserTopicsSave(userId, subTopicId) {
+    async postUserTopicsSave(userId, subTopicId, req) {
+        if(req.session.userId !== userId) return {code: 401, message: 'Unauthorized!'};
         let postUserTopicsSave = await UsersModel.postUserTopicsSave(userId, subTopicId);
         return postUserTopicsSave;
     }
 
-    async deleteUserTopicsSave(userId, subTopicId) {
+    async deleteUserTopicsSave(userId, subTopicId, req) {
+        if(req.session.userId !== userId) return {code: 401, message: 'Unauthorized!'};
         let deleteUserTopicsSave = await UsersModel.deleteUserTopicsSave(userId, subTopicId);
         return deleteUserTopicsSave;
+    }
+
+    async postUserMediaSave(userId, mediaId, req) {
+        if(req.session.userId !== userId) return {code: 401, message: 'Unauthorized!'};
+        let postUserMediaSave    =  await UsersModel.postUserMediaSave(userId, mediaId);
+
+        if(postUserMediaSave && postUserMediaSave.rowCount) {
+            return {
+                code: 200,
+                message: 'Media Saved Succesfully'
+            };
+        } else {
+            return {
+                code: 400,
+                message: 'Media Not Saved Succesfully'
+            };
+        }
+    }
+
+    async deleteUserMediaSave(userId, mediaId, req) {
+        if(req.session.userId !== userId) return {code: 401, message: 'Unauthorized!'};
+        let deleteUserMediaSave    =  await UsersModel.deleteUserMediaSave(userId, mediaId);
+
+        if(deleteUserMediaSave && deleteUserMediaSave.rowCount) {
+            return {
+                code: 200,
+                message: 'Media Deleted Succesfully'
+            };
+        } else {
+            return {
+                code: 400,
+                message: 'Media Not Deleted Succesfully'
+            };
+        }
+    }
+
+    async savedMedia(userId, req) {
+        if(req.session.userId !== userId) return {code: 401, message: 'Unauthorized!'};
+        let postUserMediaSave    =  await UsersModel.savedMedia(userId);
+
+        if(postUserMediaSave && postUserMediaSave.rowCount) {
+            return {
+                code: 200,
+                data: postUserMediaSave.rows
+            };
+        } else {
+            return {
+                code: 400
+            };
+        }
     }
 }
